Cleanup - Moved files into respective folders - Removed unwanted code and comments - Started creating main menu screen
